+++ packages/shield-controller/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Log `not_shown` if result is not available ([#6667](https://github.com/MetaMask/core/pull/6667))
+- Add `message` and `reasonCode` to coverage result type ([#6797](https://github.com/MetaMask/core/pull/6797))
 
 ### Changed